+++ GeosCore/chemistry_mod.F90
@@ -214,7 +214,6 @@ CONTAINS
          State_Chm      = State_Chm,                                         &
          State_Grid     = State_Grid,                                        &
          State_Met      = State_Met,                                         &
-         mapping        = State_Chm%Map_All,                                 &
          new_units      = KG_SPECIES,                                        &
          previous_units = previous_units,                                    &
          RC             = RC                                                )
@@ -1075,7 +1074,6 @@ CONTAINS
          State_Chm  = State_Chm,                                             &
          State_Grid = State_Grid,                                            &
          State_Met  = State_Met,                                             &
-         mapping    = State_Chm%Map_All,                                     &
          new_units  = previous_units,                                        &
          RC         = RC                                                    )