+++ GeosCore/vdiff_mod.F90
@@ -2316,7 +2316,6 @@ CONTAINS
          State_Chm      = State_Chm,                                         &
          State_Grid     = State_Grid,                                        &
          State_Met      = State_Met,                                         &
-         mapping        = State_Chm%Map_All,                                 &
          new_units      = MOLES_SPECIES_PER_MOLES_DRY_AIR,                   &
          previous_units = previous_units,                                    &
          RC             = RC                                                )
@@ -2389,7 +2388,6 @@ CONTAINS
          State_Chm  = State_Chm,                                             &
          State_Grid = State_Grid,                                            &
          State_Met  = State_Met,                                             &
-         mapping    = State_Chm%Map_All,                                     &
          new_units  = previous_units,                                        &
          RC         = RC                                                    )